Fix a few [-Wsign-compare] cases in libyottadb.h and ydb*.c
-Wsign-compare is a compiler warning flag that currently produces a thousands of warnings
and hence is currently not enabled in sr_linux/platform.cmake.

But we noticed a few of these warnings in recently introduced code (SimpleAPI). Those
cases are fixed as part of this commit.

+++ sr_unix/libyottadb.h
@@ -154,7 +154,7 @@ enum
  */
 #define YDB_COPY_STRING_TO_BUFFER(STRING, BUFFERP, COPY_DONE)	\
 {								\
-	int	len;						\
+	size_t	len;						\
 								\
 	len = strlen(STRING);					\
 	if (len <= (BUFFERP)->len_alloc)			\
